% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lambda.R
\name{calculate_lambda}
\alias{calculate_lambda}
\title{Calculate Lambda

takes contact matrix, current state of model, and parameter set and calculates the number of contacts
each age group have with infected people}
\usage{
calculate_lambda(mixing_matrix, init_pop, parms)
}
\arguments{
\item{mixing_matrix}{matrix which has the number of contacts age group i has with age group j e.g. mixing_matrix[1,2]
is the number of contacts age group 1 has with age group 2,}

\item{init_pop}{current state of the model}

\item{parms}{a list of parameters}
}
\value{
a vector which has the number of contacts a person in each age group has with an infected person each day
}
\description{
Calculate Lambda

takes contact matrix, current state of model, and parameter set and calculates the number of contacts
each age group have with infected people
}
