% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covid_19_model_function_v3.R
\name{covid_19_model_function}
\alias{covid_19_model_function}
\title{Model Function for MN covid-19 model

function contains a list of discrete differential equations
uses given parameter and timestep to calculate difference in each timestep at specified time

states are
S-sucsceptible
E-pre-infectious, E1-Em
AI- asymptomatic infectious, AI1-AIn
I-infectious, I1-In
H- person in hospital
ICU-person in ICU
R-recovered
D-deaths}
\usage{
covid_19_model_function(
  t,
  v_model_state,
  parms,
  v_days_since_peak,
  v_strat_active_tlast
)
}
\arguments{
\item{t}{current time}

\item{v_model_state}{current state of model, i.e. number of people in each state, comorbidity, and age group}

\item{parms}{list of parameters}

\item{v_days_since_peak}{vector of the number of days since the last peak in infections, hospitalizations, and deaths}
}
\value{
vector of the difference in each model state at specified time step
}
\description{
Model Function for MN covid-19 model

function contains a list of discrete differential equations
uses given parameter and timestep to calculate difference in each timestep at specified time

states are
S-sucsceptible
E-pre-infectious, E1-Em
AI- asymptomatic infectious, AI1-AIn
I-infectious, I1-In
H- person in hospital
ICU-person in ICU
R-recovered
D-deaths
}
