% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_states_index.R
\name{gen_states_index}
\alias{gen_states_index}
\title{Generate epi states and their indecies}
\usage{
gen_states_index(
  age_groups = c(0, 10, 20, 30, 40, 50, 60, 70, 80),
  comorbidity_groups = c("c0", "c1"),
  n_exposed_states,
  n_infected_states
)
}
\arguments{
\item{age_groups}{a vector of age group}

\item{comorbidity_groups}{a string vector of different comorbidity groups}

\item{n_exposed_states}{number of exposed states}

\item{n_infected_states}{number of infected states}
}
\value{
a list of fixed population characteristics for the simulation
}
\description{
Generate epi states and their indecies
}
