% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_initial_conditions.R
\name{get_initial_conditions}
\alias{get_initial_conditions}
\title{Initial condition generation}
\usage{
get_initial_conditions(parms, m_init_cases)
}
\arguments{
\item{parms}{list of parameters generated by parameters() function}

\item{m_init_cases}{data.frame of initial cases in MN

uses an rda of data from March 22nd to generate the initial conditions of the model}
}
\value{
vector which has the initial conditions of the simulation
}
\description{
Initial condition generation
}
