% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{m_init_cases}
\alias{m_init_cases}
\title{m_init_cases}
\format{
An object of class \code{data.frame} with 5 rows and 9 columns.
}
\usage{
m_init_cases
}
\description{
m_init_cases
}
\keyword{datasets}
