% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mixing_matrix}
\alias{mixing_matrix}
\title{mixing_matrix}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 9 rows and 9 columns.
}
\usage{
mixing_matrix
}
\description{
mixing_matrix
}
\keyword{datasets}
