% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lambda.R
\name{modify_mixing_matrix}
\alias{modify_mixing_matrix}
\title{Modify mixing matrix via contact reduction}
\usage{
modify_mixing_matrix(t, parms, v_days_since_peak, v_strat_active_tlast)
}
\arguments{
\item{t}{time step.}

\item{parms}{a list of parameters.}

\item{v_days_since_peak}{a vector of days since the peak of the measure specified in `str_peak_type`.}
}
\description{
Modify mixing matrix via contact reduction
}
