% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{population}
\alias{population}
\title{population}
\format{
An object of class \code{data.frame} with 101 rows and 2 columns.
}
\usage{
population
}
\description{
population
}
\keyword{datasets}
