% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_output_v3.R
\name{process_output}
\alias{process_output}
\title{Process Output

takes raw output from solve function and calculates cumulative infections, prevalent infections, daily deaths, cumulative deaths,
number of people in the hospital, number of people in the ICU for each timestep}
\usage{
process_output(out, parms)
}
\arguments{
\item{out}{matrix or list of raw output from solve function}

\item{parms}{list of parameters}
}
\value{
the same matrix that was inputted but with columns at the start for each calculated output
}
\description{
Process Output

takes raw output from solve function and calculates cumulative infections, prevalent infections, daily deaths, cumulative deaths,
number of people in the hospital, number of people in the ICU for each timestep
}
