% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solve_model.R
\name{solve_model}
\alias{solve_model}
\title{Solve Model

takes initial conditions, sequence of times, model function, 
and parameter list and calculates the state of the model at each day timestep}
\usage{
solve_model(times, func, parms, return_full = 0, return_incidence = FALSE)
}
\arguments{
\item{times}{vector of times to calculate model state at}

\item{func}{function which generates differences between model states, \code{\link{covid_19_main_function}}}

\item{parms}{list of parameters}

\item{return_full}{default = 0 for integer day output; set to 1 if returning full output matrix by time step}
}
\value{
matrix of raw results, each row is 1 timestep, each column is a different model state
}
\description{
Solve Model

takes initial conditions, sequence of times, model function, 
and parameter list and calculates the state of the model at each day timestep
}
