% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cycling_trigger.R
\name{update_end_time_trigger_strategy}
\alias{update_end_time_trigger_strategy}
\title{Updating the end time of the trigger strategy}
\usage{
update_end_time_trigger_strategy(
  tnext,
  v_days_since_peak,
  str_peak_type,
  end_day_peak_trigger,
  parms
)
}
\arguments{
\item{tnext}{Next time step}

\item{parms}{List of parameters}
}
\description{
Updating the end time of the trigger strategy
}
