% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cycling_trigger.R
\name{update_start_time_trigger_strategy}
\alias{update_start_time_trigger_strategy}
\title{Updating the start time of the trigger strategy}
\usage{
update_start_time_trigger_strategy(tnext, criteria, parms)
}
\arguments{
\item{tnext}{Next time step}

\item{criteria}{Criteria calculated beforehand}

\item{parms}{List of parameters}
}
\description{
Updating the start time of the trigger strategy
}
